+++ lib/screens/onboarding_screen.dart
@@ -1,6 +1,7 @@
 // Copyright (c) Silence Laboratories Pte. Ltd.
 // This software is licensed under the Silence Laboratories License Agreement.
 
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:lottie/lottie.dart';
@@ -59,6 +60,7 @@ class OnboardingScreen extends StatelessWidget {
                     final analyticManager = Provider.of<AnalyticManager>(context, listen: false);
                     analyticManager.trackAppStart();
                     Provider.of<AppPreferences>(context, listen: false).setIsOnboardingCompleted(true);
+                    FirebaseCrashlytics.instance.log("Onboarding done");
                   },
                   child: Row(crossAxisAlignment: CrossAxisAlignment.center, children: [
                     Text(