+++ lib/screens/sign/approve_transcation_screen.dart
@@ -2,6 +2,7 @@
 // This software is licensed under the Silence Laboratories License Agreement.
 
 import 'package:async/async.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:provider/provider.dart';
@@ -67,14 +68,14 @@ class _ApproveTransactionScreenState extends State<ApproveTransactionScreen> {
   }
 
   Future<void> _onSignature(String signature) async {
-    print("Successfully generated signature: $signature");
+    FirebaseCrashlytics.instance.log('Successfully generated signature');
     _updateTransactionState(TransactionState.signed);
     _close();
     analyticManager.trackSignPerform(status: SignPerformStatus.success);
   }
 
   void _onError(Object error) {
-    print("Error generating signature: $error");
+    FirebaseCrashlytics.instance.log('Error generating signature: $error');
     _updateTransactionState(TransactionState.failed);
     _close();
     analyticManager.trackSignPerform(status: SignPerformStatus.failed, error: error.toString());
@@ -85,7 +86,7 @@ class _ApproveTransactionScreenState extends State<ApproveTransactionScreen> {
     final appRepository = Provider.of<AppRepository>(context, listen: false);
 
     if (approved) {
-      print('Approved');
+      FirebaseCrashlytics.instance.log('Transaction approved');
       analyticManager.trackSignPerform(status: SignPerformStatus.approved);
       _updateTransactionState(TransactionState.signing);
       _signingOperation = appRepository.approve(requestModel.signRequest);
@@ -94,7 +95,7 @@ class _ApproveTransactionScreenState extends State<ApproveTransactionScreen> {
         _onError(error);
       });
     } else {
-      print('Declined');
+      FirebaseCrashlytics.instance.log('Transaction declined');
       analyticManager.trackSignPerform(status: SignPerformStatus.rejected);
       _updateTransactionState(TransactionState.canceled);
       appRepository.decline(requestModel.signRequest);